// src/pages/listar-operacoes/Interfaces.ts
export interface StateListarOperacoes {
    etapaOperacao: string;
    situacaoNegociacao: string;
    formaPagamento: string;
    operacaoCliente: string;
    documentoCliente: string;
    documentoEmpresa: string;
    nomeCliente: string;
    nomeEmpresa: string;
}

export interface StateListaOperacoes {
    operacaoCliente: string;
    documentoCliente: string;
    documentoEmpresa: string;
    nomeCliente: string;
    nomeEmpresa: string;
}